Added detection drives other than C:\
The script now iterates through a list of letters in case the user has VS installed in a drive other than C:\, to prevent confusion and manual editing of the script. In case no installation was found in the default Program Files (x86) folder, the user is prompted to enter the path manually.